Promote LabelBase as FormLabel. Increase RadioGroup blur delay for better label focus experience when jumping between radios.
