+++ src/Form/FormLabel.js
@@ -3,11 +3,20 @@
 import React, { PropTypes } from 'react';
 import { createStyleSheet } from 'jss-theme-reactor';
 import classNames from 'classnames';
-import LabelBase from '../internal/LabelBase';
 
-export const styleSheet = createStyleSheet('FormLabel', () => {
+export const styleSheet = createStyleSheet('FormLabel', (theme) => {
+  const focusColor = theme.palette.accent.A200;
   return {
-    root: {}, // expected to be used
+    root: {
+      color: theme.palette.text.secondary,
+      lineHeight: 1,
+    },
+    focused: {
+      color: focusColor,
+    },
+    error: {
+      color: theme.palette.error[500],
+    },
   };
 });
 
@@ -15,16 +24,31 @@ export default function FormLabel(props, context) {
   const {
     children,
     className: classNameProp,
+    error,
+    focused,
+    required,
     ...other
   } = props;
   const classes = context.styleManager.render(styleSheet);
 
-  const className = classNames(classes.root, classNameProp);
+  const className = classNames(classes.root, {
+    [classes.focused]: focused,
+    [classes.error]: error,
+  }, classNameProp);
+
+  const asteriskClassName = classNames({
+    [classes.error]: error,
+  });
 
   return (
-    <LabelBase className={className} {...other}>
+    <label className={className} {...other}>
       {children}
-    </LabelBase>
+      {required && (
+        <span className={asteriskClassName} data-mui-test="FormLabelAsterisk">
+          {'\u2009*'}
+        </span>
+      )}
+    </label>
   );
 }
 
@@ -42,22 +66,22 @@ FormLabel.propTypes = {
    */
   error: PropTypes.bool,
   /**
-   * Whether the inputs this label refers to are focused.
+   * Whether the input of this label is focused (used by `Group` components).
    */
   focused: PropTypes.bool,
   /**
-   * Whether this label should indicate that a response is required.
+   * Whether this label should indicate that the input.
+   * is required.
    */
   required: PropTypes.bool,
 };
 
 FormLabel.defaultProps = {
-  error: false,
+  focused: false,
   required: false,
+  error: false,
 };
 
 FormLabel.contextTypes = {
   styleManager: PropTypes.object.isRequired,
 };
-
-FormLabel.muiName = 'FormLabel';
